+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.item;
 
-import ru.practicum.shareit.dto.ItemDto;
-import ru.practicum.shareit.dto.ItemWithBookingDto;
+import ru.practicum.shareit.item.dto.ItemDto;
+import ru.practicum.shareit.item.dto.ItemWithBookingDto;
 
 import java.util.List;
