+++ src/main/java/ru/practicum/shareit/item/dto/ItemWithBookingDto.java
@@ -1,22 +1,21 @@
-package ru.practicum.shareit.dto;
+package ru.practicum.shareit.item.dto;
 
 import lombok.*;
 import ru.practicum.shareit.Create;
 import ru.practicum.shareit.Update;
+import ru.practicum.shareit.item.comment.dto.CommentDto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.util.List;
 
-/**
- * TODO Sprint add-controllers.
- */
 @Getter
 @Setter
 @ToString
 @EqualsAndHashCode(of = "id")
 @NoArgsConstructor
 @AllArgsConstructor
-public class ItemDto {
+public class ItemWithBookingDto {
 
     @NotNull(groups = {Update.class})
     private Long id;
@@ -27,4 +26,17 @@ public class ItemDto {
     @NotNull(groups = {Create.class})
     private Boolean available;
     private Long requestId;
+    private List<CommentDto> comments;
+    private Booking lastBooking;
+    private Booking nextBooking;
+
+    @Getter
+    @Setter
+    @EqualsAndHashCode(of = "id")
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class Booking {
+        Long id;
+        Long bookerId;
+    }
 }
