+++ src/main/java/ru/practicum/shareit/item/model/ItemMapper.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.item.comment.CommentMapper;
-import ru.practicum.shareit.dto.ItemDto;
-import ru.practicum.shareit.dto.ItemWithBookingDto;
+import ru.practicum.shareit.item.dto.ItemDto;
+import ru.practicum.shareit.item.dto.ItemWithBookingDto;
 import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
