+++ src/main/resources/schema.sql
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS items
     id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
     name VARCHAR(255) NOT NULL,
     description VARCHAR(1000) NOT NULL,
-    available BOOLEAN,
+    is_available BOOLEAN,
     owner_id BIGINT NOT NULL,
     request_id BIGINT,
     CONSTRAINT pk_item PRIMARY KEY (id),
